LUOGU

+++ P1102 A-B 数对.cpp
@@ -5,8 +5,8 @@
 #include<algorithm>
 using namespace std;
 int main(){
-	int n,c,sum=0;
-	int a[2020];
+	 int n,c,sum=0;
+	long long a[2020];
 	scanf("%d%d",&n,&c);
 	map<int,int> m;
 	for(int i=0;i<n;i++){
@@ -16,6 +16,6 @@ int main(){
 	for(int i=0;i<n;i++){
 		sum+=m[a[i]+c];
 	}
-	printf("%d",sum);
+	printf("%lld",sum);
 	return 0;
 }
